client, store: don't use store from client (use client from store)
This moves the definition of BuyOptions and BuyResult from store to
client, and inverts the dependency.

The dependency graph goes from

![dependency graph before the refactor](https://r.chipaca.com/client_before.png)

to

![dependency graph after the refactor](https://r.chipaca.com/client_after.png)
